yfinance optimize
- Optimizing adding start and end dates, and threading, download multiple stocks all at once.

Downloading reduces time about 2 seconds for from 11.5 to 9.5 on average for one year backtest with five tickers.

When downloading in bulk, not using max, yf will return dates not including the last. To compensate, seven days were added and then refiltered.

For the historical datetime index, the latest index will be pickled locally, and retrieved for subsequent test, and saved each test. This will illiminate the need for a max download with all ohlc for each backtest. Also dropped the ohlcv info right away and managed the datetimes in a list/set for performance.