Catching backtest positional arguments error, strategy name and budget.
Error being thrown when user uses the old method of passing in the strategy name and budget as positional arguments.

This results in an AttributeError for the backtesting_start date.

This commit catches the error and returns a message instructing the user to change the `name` and `budget` positional arguments to keyword arguements.