Have alpaca class find and track the correct positions based on strategy name in orders.
Alpaca now saves the strategy name as part of the client_order_id, which makes it possible to
parse all orders that made up a position and decide if the position was opened by our strategy
or not. If it was we want to track it, if not we'll ignore it.