Futures quarterly contract rollover logic fix
Each quarter the index futures roll over on the 3rd Friday of the month (March, June, Sept, Dec). 
Mid way through the month it will automatically roll instead of waiting until the end of the month which exposes the a couple of weeks of an expired contract. 
TODO: other types of futures might have different calendars to the main index futures such as ES, NQ, etc..