Set live parameter to False in projectx_data.py
Last week's projectx changes flipped the retrieveBars call's "live" parameter from False to True. Can we change it back because the API call errors out and doesn't return any historical bars for me. The ProjectX just defines that param as "Whether to retrieve bars using the sim or live data subscription." I can't find any other info on it. If it's ok to set back to False here's the PR .